+++ jabgui/src/main/java/org/jabref/gui/JabRefGUI.java
@@ -48,6 +48,7 @@ import org.jabref.logic.net.ProxyRegisterer;
 import org.jabref.logic.os.OS;
 import org.jabref.logic.protectedterms.ProtectedTermsLoader;
 import org.jabref.logic.remote.RemotePreferences;
+import org.jabref.logic.remote.server.ConnectorTokenManager;
 import org.jabref.logic.remote.server.RemoteListenerServerManager;
 import org.jabref.logic.search.IndexManager;
 import org.jabref.logic.search.PostgreServer;
@@ -94,6 +95,7 @@ public class JabRefGUI extends Application {
 
     private static RemoteListenerServerManager remoteListenerServerManager;
     private static HttpServerManager httpServerManager;
+    private static ConnectorTokenManager connectorTokenManager;
     private static LanguageServerController languageServerController;
 
     private Stage mainStage;
@@ -186,7 +188,11 @@ public class JabRefGUI extends Application {
         JabRefGUI.remoteListenerServerManager = new RemoteListenerServerManager();
         Injector.setModelOrService(RemoteListenerServerManager.class, JabRefGUI.remoteListenerServerManager);
 
+        JabRefGUI.connectorTokenManager = new ConnectorTokenManager(preferences.getRemotePreferences());
+        Injector.setModelOrService(ConnectorTokenManager.class, JabRefGUI.connectorTokenManager);
+
         JabRefGUI.httpServerManager = new HttpServerManager();
+        JabRefGUI.httpServerManager.setTokenManager(JabRefGUI.connectorTokenManager);
         Injector.setModelOrService(HttpServerManager.class, JabRefGUI.httpServerManager);
 
         JabRefGUI.languageServerController = new LanguageServerController(preferences, journalAbbreviationRepository, entryTypesManager);