+++ jablib/src/main/java/org/jabref/logic/preferences/JabRefCliPreferences.java
@@ -369,6 +369,7 @@ public class JabRefCliPreferences implements CliPreferences {
     private static final String ENABLE_LANGUAGE_SERVER = "enableLanguageServer";
     private static final String LANGUAGE_SERVER_PORT = "languageServerPort";
     private static final String ALLOWED_ORIGINS = "allowedOrigins";
+    private static final String API_TOKEN = "apiToken";
 
     private static final String AI_ENABLED = "aiEnabled";
     private static final String AI_AUTO_GENERATE_EMBEDDINGS = "aiAutoGenerateEmbeddings";
@@ -645,6 +646,7 @@ public class JabRefCliPreferences implements CliPreferences {
         defaults.put(ENABLE_LANGUAGE_SERVER, Boolean.FALSE);
         defaults.put(LANGUAGE_SERVER_PORT, 2087);
         defaults.put(ALLOWED_ORIGINS, "chrome-extension://;moz-extension://;https://jabref.github.io;https://jabref.org");
+        defaults.put(API_TOKEN, "");
 
         defaults.put(EXTERNAL_JOURNAL_LISTS, "");
         defaults.put(USE_AMS_FJOURNAL, true);
@@ -1347,7 +1349,8 @@ public class JabRefCliPreferences implements CliPreferences {
                 getBoolean(ENABLE_HTTP_SERVER),
                 getBoolean(ENABLE_LANGUAGE_SERVER),
                 getInt(LANGUAGE_SERVER_PORT),
-                getStringList(ALLOWED_ORIGINS));
+                getStringList(ALLOWED_ORIGINS),
+                get(API_TOKEN));
 
         EasyBind.listen(remotePreferences.portProperty(), (_, _, newValue) -> putInt(REMOTE_SERVER_PORT, newValue));
         EasyBind.listen(remotePreferences.useRemoteServerProperty(), (_, _, newValue) -> putBoolean(USE_REMOTE_SERVER, newValue));
@@ -1355,6 +1358,7 @@ public class JabRefCliPreferences implements CliPreferences {
         EasyBind.listen(remotePreferences.enableHttpServerProperty(), (_, _, newValue) -> putBoolean(ENABLE_HTTP_SERVER, newValue));
         EasyBind.listen(remotePreferences.languageServerPortProperty(), (_, _, newValue) -> putInt(LANGUAGE_SERVER_PORT, newValue));
         EasyBind.listen(remotePreferences.enableLanguageServerProperty(), (_, _, newValue) -> putBoolean(ENABLE_LANGUAGE_SERVER, newValue));
+        EasyBind.listen(remotePreferences.apiTokenProperty(), (_, _, newValue) -> put(API_TOKEN, newValue));
         remotePreferences.getAllowedOrigins().addListener((InvalidationListener) _ ->
                 putStringList(ALLOWED_ORIGINS, remotePreferences.getAllowedOrigins()));