+++ jablib/src/main/java/org/jabref/logic/remote/RemotePreferences.java
@@ -10,6 +10,8 @@ import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleBooleanProperty;
 import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
@@ -32,8 +34,9 @@ public class RemotePreferences {
     private final IntegerProperty languageServerPort;
 
     private final ObservableList<String> allowedOrigins;
+    private final StringProperty apiToken;
 
-    public RemotePreferences(int port, boolean useRemoteServer, int httpPort, boolean enableHttpServer, boolean enableLanguageServer, int languageServerPort, List<String> allowedOrigins) {
+    public RemotePreferences(int port, boolean useRemoteServer, int httpPort, boolean enableHttpServer, boolean enableLanguageServer, int languageServerPort, List<String> allowedOrigins, String apiToken) {
         this.port = new SimpleIntegerProperty(port);
         this.useRemoteServer = new SimpleBooleanProperty(useRemoteServer);
         this.httpPort = new SimpleIntegerProperty(httpPort);
@@ -41,6 +44,7 @@ public class RemotePreferences {
         this.enableLanguageServer = new SimpleBooleanProperty(enableLanguageServer);
         this.languageServerPort = new SimpleIntegerProperty(languageServerPort);
         this.allowedOrigins = FXCollections.observableArrayList(allowedOrigins);
+        this.apiToken = new SimpleStringProperty(apiToken);
     }
 
     public int getPort() {
@@ -133,6 +137,18 @@ public class RemotePreferences {
 
     public void setAllowedOrigins(List<String> allowedOrigins) {
         this.allowedOrigins.setAll(allowedOrigins);
+    }
+
+    public String getApiToken() {
+        return apiToken.getValue();
+    }
+
+    public StringProperty apiTokenProperty() {
+        return apiToken;
+    }
+
+    public void setApiToken(String apiToken) {
+        this.apiToken.setValue(apiToken);
     }
 
     /// Gets the IP address where both the remote server and the http server are listening.