+++ jablib/src/test/java/org/jabref/logic/remote/RemotePreferencesTest.java
@@ -15,7 +15,7 @@ class RemotePreferencesTest {
 
     @BeforeEach
     void setUp() {
-        preferences = new RemotePreferences(1000, true, 3000, false, false, 2087, List.of());
+        preferences = new RemotePreferences(1000, true, 3000, false, false, 2087, List.of(), "");
     }
 
     @Test